Adding composer.json, moving Module class to PSR-0 compliant path, adding reference to the asset manager for loading files within 'public/'
