+++ src/KJSencha/Module.php
@@ -1,0 +1,40 @@
+<?php
+
+namespace KJSencha;
+
+use Zend\ModuleManager\Feature\AutoloaderProviderInterface;
+use Zend\ModuleManager\Feature\ConfigProviderInterface;
+use Zend\ModuleManager\Feature\ServiceProviderInterface;
+
+class Module implements AutoloaderProviderInterface, ConfigProviderInterface, ServiceProviderInterface
+{
+    /**
+     * {@inheritDoc}
+     */
+    public function getAutoloaderConfig()
+    {
+        return array(
+            'Zend\Loader\StandardAutoloader' => array(
+                'namespaces' => array(
+                    __NAMESPACE__ => __DIR__,
+                ),
+            ),
+        );
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function getConfig()
+    {
+        return include __DIR__ . '/../../config/module.config.php';
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function getServiceConfig()
+    {
+        return include __DIR__ . '/../../config/services.config.php';
+    }
+}
